requirements.txt
flask
faunadb
python-dotenv
pytz
twofish
base64
